% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_name.R
\name{make_full_name}
\alias{make_full_name}
\title{Create vector of full names}
\usage{
make_full_name(x, ...)
}
\arguments{
\item{x}{Input should be either a data frame with name information or a
character vector of first names and additional character vectors of middle
names, last names, and suffix. If a data frame is provided, it should at a
minimum contain the variables 'first_name' and 'last_name' but will also
used columns named 'middle_name' or 'suffix' if found in the data frame
as well.}

\item{...}{Additional character vectors only used if x is a vector of first
names.}
}
\value{
The output will match the input–so if data frame then out data frame.
}
\description{
Converts first_name, middle_name, last_name, suffix into a single, full-name
character vector
}
